Check the magnitude of integer types in validate().
This partially fixes Github issue #131.

This is not a complete solution, as the error message is still opaque,
but it's an improvement.  Throwing the error at struct creation time
instead of at serialization time makes it clearer which struct has a
bad field, and makes it easier for TChannel clients to handle, since
TChannel doesn't seem to provide a way to catch response serialization
exceptions.