setup.py: Don't check cython version
In setup.py, we verify that we are using at least Cython 0.17. This
check is causing failures if the version of Cython being used is a pre
or post-release (0.20.1post0, for example).

It's unclear why the check was there versus in the requirements.txt in
the first place. Further, the version of Cython is relevant only at
development and release time.  Since it's about 4 years later and Cython
is on release 0.28, it's pretty safe to assume that we won't run into
0.17 on machines we release ThriftRW from.

This change drops the check from the setup.py and transfers the
constraint to requirements.txt.